Moved upgrade function out to a new file.
